requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
Pillow>=10.0.1
PyMuPDF>=1.22.2
python-dotenv>=0.19.0
pikepdf>=8.0.0
supabase>=2.0.0
aiofiles>=23.0.0
redis[asyncio]>=5.0.0
aiohttp>=3.9.0